fix: add persistence for main (#73)
fix: add persistence for plant data

fix: fix index out range error

fix: correct persistence logic

fix: add persistence for progress bar

fix: add persistence for tap button status

fix: add persistence for dialgoue data

+++ ForestTori/ForestTori/Source/View/Garden/GardenView.swift
@@ -20,7 +20,7 @@ struct GardenView: View {
     
     private let noPlantCaption = "아직 다 키운 식물이 없어요."
     private let summerMessage = "여름 하늘은 봄보다 더 높아져서 더 멀리까지 바라볼 수 있는 거 알아?"
-    var totalProgressValue: Float?
+    var totalProgressValue: Double?
     
     var body: some View {
         NavigationView {
